fix: document and test the descriptor address-form contract
Review of the address wrappers found the headers silent on what a caller has
to guarantee, and two forwarding paths with no test coverage.

Document the contract once per binding block rather than per declaration. The
trap worth naming is a caller passing a batch's full column count while handing
over only the selected columns: the output length still matches what the kernel
expects, so the length check passes and the kernel reads past the end of the
input. The blocks also record that the buffers must stay pinned across the call
because the bindings drop the GIL, and that the unconditional zero-address
rejection is deliberately stricter than the pointer forms it delegates to. The
four per-parameter comments on PDistNumericAddress are gone; one authoritative
statement beats two that can drift, and the surrounding headers use a one-line
brief for these helpers.

Add two tests for forwarding that nothing previously exercised. Every cdist
fixture passed the same buffer, row count and mask for both sides, so swapping
a and b, or dropping the missing policy, changed nothing observable;
CDistNumericIntoAddress was never called directly at all. The new fixture
slices one 5x2 buffer into two unequal row ranges with different masks, which
keeps a swapped read in bounds so a mistake shows up as a wrong value rather
than as undefined behaviour. The statistics forms were only ever called with a
zero validity address and a default rcond, so neither was forwarded under test;
they now are.

Release the GIL for the three batch statistics overloads. Including
descriptor_statistics.h wraps more than the address forms, and the batch
overloads are an O(rows x columns) pass -- a full eigendecomposition for the
inverse -- that was holding the GIL throughout.

Convert the pointer-to-address casts to the form used at the 21 pre-existing
sites in this repo. A direct reinterpret_cast to a wider integer type is
ill-formed where uint64_t is wider than a pointer.

+++ src/descriptor_compare.cpp
@@ -782,7 +782,9 @@ std::vector<double> PDistNumericAddress(
     }
 
     PDistNumericIntoAddress(values_address, validity_address, rows, columns, metric, missing,
-                            reinterpret_cast<std::uint64_t>(output.data()), output.size(),
+                            static_cast<std::uint64_t>(
+                                reinterpret_cast<std::uintptr_t>(output.data())),
+                            output.size(),
                             kernel);
     return output;
 }
@@ -836,7 +838,9 @@ std::vector<double> CDistNumericAddress(
 
     CDistNumericIntoAddress(a_values_address, a_validity_address, a_rows, b_values_address,
                             b_validity_address, b_rows, columns, metric, missing,
-                            reinterpret_cast<std::uint64_t>(output.data()), output.size(),
+                            static_cast<std::uint64_t>(
+                                reinterpret_cast<std::uintptr_t>(output.data())),
+                            output.size(),
                             kernel);
     return output;
 }
